pyproject.toml
[project]
name = "kafka-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.2.15",
]
